Added "Cycle Test" function and ability to change the motor direction in the software.
The Cycle Test runs the head down until it arcs, then returns it to the
starting position.  This is repeated until the user exits the test. The
depth counter can be cleared before starting the test (using the Jog Mode
menu).  The counter should return to nearly the same reading each touch
down, increasing slowly as a cut is made.  Passing this test virtually
guarantees that the device is working properly.

The motor direction can now be switched by the user.  This allows heads
and cables with reversed wiring to be used.  The setting is saved in the
EEprom.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>