Removed extra motor up command from cycleTest which was not getting counted by the retract counter.  In the cutNotch function, this extra move was used to retract as quickly as possible.  In cycleTest, this is not as critical so it could be deleted.
This bug was causing the head to retract past the original position at the
top of the stroke more and more with each cycle.

The second call to move the motor up had to be changed from
pulseMotorWithDelay to pulseMotorUpWithDelay because the former requires
the latter to be called first to set the motor direction.


Signed-off-by: Mike Schoonover <mike@folsomvillage.com>